Day 7, continued sabotage with flawed equipment
The elves have now given me a computer with only 70MB of space, and it's
completely full of random junk. They've told me to run updates on it,
but it needs 30MB free to do so, and it only responds to `cd` and `ls`.
Why the hell does this thing not have `du`?

To make matters worse, it doesn't seem to have a way to delete a single
file, only a whole directory, and they insist that every directory
that's not the smallest directory that's big enough to free up enough
space is "essential to the functioning of the christmas spirit".

Which, of course, means they know exactly which directory it is I need
to delete. They just won't tell me. I suspect this is why they live in
isolation at the North Pole - no one will tolerate them anywhere else.