Day 8, elves are also doing nothing about climate change
A previous expedition planted a bunch of trees as part of a
reforestation effort, nice work! Surely we'll do the same, leave the
world better than we found it, right?

No, we're going to cut some to get wood to build a tree house in the
most optimal position with a maximal "scenic score".

At least I'll sleep in a dry place tonight, for a change.