Day 9, elves tried to murder me with a bridge
I was already suspicious of this rickety looking bridge to begin with,
so when the elves started rambling something about planck lengths making
it stable and easy to prove, I figured I had to go test that theory.

So, turns out it's not too bad to simulate... so long as you believe
them. I should've known better by now, and their simulation was flawed.
Ropes were tensioned like mad too, so instead of trying to grab them I
had to avoid them.

Every day an adventure.