Day 16... and what a day
So, turns out there were elephants at the distress signal!
Unfortunately, it also turns out this is a volcano!

Naturally, your average elephant is far smarter than your top of the
line elf, so I was able to teach one to activate the pressure relief
valves in pretty short order and try to delay an inevitable eruption.
Make for the exits!